Minions/requirements.txt
python-dotenv
openai
google-genai
requests
SkillLink
AgentToAgent
